Ground reward shop accounting foundation

+++ backend/src/main/java/coin/coinzzickmock/feature/account/api/AccountSummaryResponse.java
@@ -8,6 +8,6 @@ public record AccountSummaryResponse(
         double available,
         double totalUnrealizedPnl,
         double roi,
-        double rewardPoint
+        int rewardPoint
 ) {
 }
